+++ backend/src/neuroagent/app/dependencies.py
@@ -79,6 +79,7 @@ from neuroagent.tools import (
     StrainGetOneTool,
     SubjectGetAllTool,
     SubjectGetOneTool,
+    WeatherTool,
     WebSearchTool,
 )
 from neuroagent.tools.base_tool import BaseTool
@@ -341,7 +342,7 @@ def get_tool_list(
         SubjectGetAllTool,
         SubjectGetOneTool,
         # NowTool,
-        # WeatherTool,
+        WeatherTool,
         # RandomPlotGeneratorTool,
     ]