+++ frontend/src/components/chat/chat-message-human.tsx
@@ -8,7 +8,7 @@ export const ChatMessageHuman = function ChatMessageHuman({
   content,
 }: ChatMessageHumanProps) {
   return (
-    <div className="flex justify-end break-all border-solid p-8">
+    <div className="flex justify-end break-all border-solid p-4 pt-6">
       <Card className="max-w-[70%]">
         <CardContent>
           <h1 className="whitespace-pre-wrap pt-8 text-lg">{content}</h1>