+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reasoning component.
 - Rate limit endpoint.
 - More entity core endpoints.
+- Rules.
 
 ### Changed
 - Entitycore autogen 2025.6.8